+++ src/lib.rs
@@ -90,3 +90,19 @@ pub mod optional;
 pub mod sensors;
 mod status;
 pub use status::*;
+
+/// Return own crate version. Used in API responses.
+pub fn get_version() -> &'static str {
+    env!("CARGO_PKG_VERSION")
+}
+
+#[cfg(test)]
+mod test {
+    use super::get_version;
+
+    #[test]
+    fn test_get_version() {
+        let version = get_version();
+        assert_eq!(3, version.split('.').count());
+    }
+}